+++ sdk/monitor/monitor-opentelemetry-exporter/src/Declarations/Constants.ts
@@ -109,7 +109,7 @@ export const ENV_AZURE_MONITOR_PREFIX = "AZURE_MONITOR_PREFIX";
 export const ENV_AZURE_MONITOR_DISTRO_VERSION = "AZURE_MONITOR_DISTRO_VERSION";
 
 /**
- * Enables the preview version of customer-facing Statsbeat.
+ * Enables the preview version of customer-facing SDK Stats.
  * @internal
  */
 export const ENV_APPLICATIONINSIGHTS_STATSBEAT_ENABLED_PREVIEW =